Move some common functionality into a shared package
