+++ common/common/__init__.py
@@ -1,0 +1,19 @@
+"""Shared schema code!"""
+from .schemas import (
+    ActionNoTargetRequest,
+    ActionWithTargetRequest,
+    ChatMessage,
+    InitializePlayer,
+    PlayerSchema,
+    RegistrationSuccessful,
+)
+
+__all__ = (
+    # schemas.py
+    "ActionNoTargetRequest",
+    "ActionWithTargetRequest",
+    "ChatMessage",
+    "InitializePlayer",
+    "PlayerSchema",
+    "RegistrationSuccessful",
+)
