+++ common/pyproject.toml
@@ -1,17 +1,12 @@
 [tool.poetry]
-name = "server"
+name = "common"
 version = "0.1.0"
 description = ""
-authors = ["AbooMinister25 <aboominister@gmail.com>"]
+authors = ["A5rocks <git@helvetica.moe>"]
 
 [tool.poetry.dependencies]
 python = "^3.10"
-pycryptodome = "^3.15.0"
-cryptography = "^37.0.4"
-websockets = "^10.3"
 pydantic = "^1.9.1"
-# "develop" just means that no need to reinstall deps to get changes
-common = { path = "../common/", develop = true }
 
 [tool.poetry.dev-dependencies]
 # Base tools
